+++ common/src/main/java/com/slymask3/instantblocks/block/instant/InstantStatueBlock.java
@@ -191,14 +191,14 @@ public class InstantStatueBlock extends InstantBlock implements EntityBlock {
 
 			Single.setup(builder,world,x,y,z).setBlock(Blocks.AIR).queue();
 
-			buildHead(builder, world, x, y, z, img, direction, blockEntity.head, blockEntity.rgb);
-			buildBody(builder, world, x, y, z, img, direction, blockEntity.body, blockEntity.rgb);
-			buildLegs(builder, world, x, y, z, img, direction, blockEntity.legLeft, blockEntity.legRight, blockEntity.rgb);
+			buildHead(builder, img, direction, blockEntity.head, blockEntity.rgb);
+			buildBody(builder, img, direction, blockEntity.body, blockEntity.rgb);
+			buildLegs(builder, img, direction, blockEntity.legLeft, blockEntity.legRight, blockEntity.rgb);
 
 			if(skin.isSlim()) {
-				buildArmsSlim(builder, world, x, y, z, img, direction, blockEntity.armLeft, blockEntity.armRight, blockEntity.rgb);
+				buildArmsSlim(builder, img, direction, blockEntity.armLeft, blockEntity.armRight, blockEntity.rgb);
 			} else {
-				buildArms(builder, world, x, y, z, img, direction, blockEntity.armLeft, blockEntity.armRight, blockEntity.rgb);
+				buildArms(builder, img, direction, blockEntity.armLeft, blockEntity.armRight, blockEntity.rgb);
 			}
 
 			builder.build();
@@ -211,374 +211,141 @@ public class InstantStatueBlock extends InstantBlock implements EntityBlock {
 		return false;
 	}
 
-	private static void buildHead(Builder builder, Level world, int x, int y, int z, BufferedImage img, Direction direction, boolean build, boolean rgb) {
+	private static void buildHead(Builder builder, BufferedImage img, Direction direction, boolean build, boolean rgb) {
 		if(build) {
-			//TOP
-			for(int xtimes=0; xtimes<8; xtimes++) {
-				for(int ytimes=0; ytimes<8; ytimes++) {
-					Single.setup(builder,world,x,y+31,z).offset(direction,-5+ytimes,-4+xtimes).setImageColor(img,8+xtimes, ytimes,rgb).queue();
-				}
-			}
-
-			//BOTTOM
-			for(int xtimes=0; xtimes<8; xtimes++) {
-				for(int ytimes=0; ytimes<8; ytimes++) {
-					Single.setup(builder,world,x,y+24,z).offset(direction,-5+ytimes,-4+xtimes).setImageColor(img,16+xtimes, ytimes,rgb).queue();
-				}
-			}
-
-			//RIGHT
-			for(int xtimes=0; xtimes<8; xtimes++) {
-				for(int ytimes=0; ytimes<8; ytimes++) {
-					Single.setup(builder,world,x,y+31-ytimes,z).offset(direction,-5+xtimes,-4).setImageColor(img, xtimes,8+ytimes,rgb).queue();
-				}
-			}
-
-			//LEFT
-			for(int xtimes=0; xtimes<8; xtimes++) {
-				for(int ytimes=0; ytimes<8; ytimes++) {
-					Single.setup(builder,world,x,y+31-ytimes,z).offset(direction, 2 -xtimes,+3).setImageColor(img,16+xtimes,8+ytimes,rgb).queue();
-				}
-			}
-
-			//FRONT
-			for(int xtimes=0; xtimes<8; xtimes++) {
-				for(int ytimes=0; ytimes<8; ytimes++) { //face is 8x8, so 8 total times
-					Single.setup(builder,world,x,y+31-ytimes,z).offset(direction,+2,-4+xtimes).setImageColor(img,8+xtimes,8+ytimes,rgb).queue(); //face starts at 8, 8
-				} //face starts x-4, y+31, z+2
-			}
-
-			//BACK
-			for(int xtimes=0; xtimes<8; xtimes++) {
-				for(int ytimes=0; ytimes<8; ytimes++) {
-					Single.setup(builder,world,x,y+31-ytimes,z).offset(direction,-5, 3 -xtimes).setImageColor(img,24+xtimes,8+ytimes,rgb).queue();
-				}
-			}
+			buildFromImage(builder,direction,rgb,5,0,3,0,31,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,8,0,8,8); //top
+			buildFromImage(builder,direction,rgb,5,0,3,0,24,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,16,0,8,8); //bottom
+			buildFromImage(builder,direction,rgb,5,0,3,0,31,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,0,8,8,8); //left
+			buildFromImage(builder,direction,rgb,0,2,0,4,31,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,16,8,8,8); //right
+			buildFromImage(builder,direction,rgb,0,2,3,0,31,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,8,8,8,8); //front
+			buildFromImage(builder,direction,rgb,5,0,0,4,31,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,24,8,8,8); //back
 		}
 	}
 
-	private static void buildBody(Builder builder, Level world, int x, int y, int z, BufferedImage img, Direction direction, boolean build, boolean rgb) {
+	private static void buildBody(Builder builder, BufferedImage img, Direction direction, boolean build, boolean rgb) {
 		if(build) {
-			//TOP
-			for(int xtimes=0; xtimes<8; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //topbody is 8x4
-					Single.setup(builder,world,x,y+23,z).offset(direction,-3+ytimes,-4+xtimes).setImageColor(img,20+xtimes,16+ytimes,rgb).queue(); //topbody starts at 20, 16
-				} //topbody starts at x-4, y+23, z+3
-			}
-
-			//BOTTOM
-			for(int xtimes=0; xtimes<8; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //botbody is 8x4
-					Single.setup(builder,world,x,y+12,z).offset(direction,-ytimes,-4+xtimes).setImageColor(img,28+xtimes,16+ytimes,rgb).queue(); //botbody starts at 28, 16
-				} //botbody starts at x-4, y+12, z
-			}
-
-			//RIGHT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightbody is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-3+xtimes,-4).setImageColor(img,16+xtimes,20+ytimes,rgb).queue(); //rightbody starts at 16, 20
-				} //rightbody starts at x-4, y+23, z-3
-			}
-
-			//LEFT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftbody is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-xtimes,+3).setImageColor(img,28+xtimes,20+ytimes,rgb).queue(); //leftbody starts at 28, 20
-				} //rightbody starts at x+3, y+23, z
-			}
-
-			//FRONT
-			for(int xtimes=0; xtimes<8; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //body is 8x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,0,-4+xtimes).setImageColor(img,20+xtimes,20+ytimes,rgb).queue(); //body starts at 20, 20
-				} //body starts at x-4, y+23, z
-			}
-
-			//BACK
-			for(int xtimes=0; xtimes<8; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //back is 8x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-3, 3 -xtimes).setImageColor(img,32+xtimes,20+ytimes,rgb).queue(); //back starts at 32, 20
-				} //back starts at x+3, y+23, z-3
-			}
+			buildFromImage(builder,direction,rgb,3,0,3,0,23,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,20, 16,8,4); //top
+			buildFromImage(builder,direction,rgb,3,0,3,0,12,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,28, 16,8,4); //bottom
+			buildFromImage(builder,direction,rgb,3,0,3,0,23,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,16, 20,4,12); //left
+			buildFromImage(builder,direction,rgb,0,0,0,4,23,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,28, 20,4,12); //right
+			buildFromImage(builder,direction,rgb,0,0,3,0,23,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,20, 20,8,12); //front
+			buildFromImage(builder,direction,rgb,3,0,0,4,23,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,32, 20,8,12); //back
 		}
 	}
 
-	private static void buildArms(Builder builder, Level world, int x, int y, int z, BufferedImage img, Direction direction, boolean buildLeft, boolean buildRight, boolean rgb) {
+	private static void buildArms(Builder builder, BufferedImage img, Direction direction, boolean buildLeft, boolean buildRight, boolean rgb) {
 		if(buildLeft) {
-			//LEFT ARM
-
-			//LEFT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightarmleft is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-xtimes,+4).setImageColor(img,48+xtimes,20+ytimes,rgb).queue(); //rightarmleft starts at 48, 20
-				} //rightarmleft starts
-			}
-
-			//RIGHT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightarmright is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-3+xtimes,+7).setImageColor(img,40+xtimes,20+ytimes,rgb).queue(); //rightarmright starts at 40, 20
-				} //rightarmright starts
-			}
-		
-			//TOP
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //shoulderleft is 4x4
-					Single.setup(builder,world,x,y+23,z).offset(direction,-3+ytimes, 4 +xtimes).setImageColor(img,44+xtimes,16+ytimes,rgb).queue(); //shoulderleft starts at 44, 16
-				} //shoulderleft starts
-			}
-	
-			//BOTTOM
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //fistleft is 4x4
-					Single.setup(builder,world,x,y+12,z).offset(direction,-3+ytimes, 4 +xtimes).setImageColor(img,48+xtimes,16+ytimes,rgb).queue(); //fistleft starts at 44, 16
-				} //fistleft starts
-			}
-			
-			//FRONT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightarmfront is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,0, 7 -xtimes).setImageColor(img,44+xtimes,20+ytimes,rgb).queue(); //rightarmfront starts at 44, 20 DONE?
-				} //rightarmfront starts
-			}
-
-			//BACK
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightarmback is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-3, 4 +xtimes).setImageColor(img,52+xtimes,20+ytimes,rgb).queue(); //rightarmback starts at 52, 20
-				} //rightarmback starts 
+			if(img.getHeight() == 32) {
+				buildFromImage(builder,direction,rgb,3,0,0,8,23,0, RelativeDirection.LEFT, RelativeDirection.BACK,img,44, 16,4,4); //top
+				buildFromImage(builder,direction,rgb,3,0,0,8,12,0, RelativeDirection.LEFT, RelativeDirection.BACK,img,48, 16,4,4); //bottom
+				buildFromImage(builder,direction,rgb,0,0,0,5,23,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,48, 20,4,12); //left
+				buildFromImage(builder,direction,rgb,3,0,0,8,23,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,40, 20,4,12); //right
+				buildFromImage(builder,direction,rgb,0,0,0,8,23,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,44, 20,4,12); //front
+				buildFromImage(builder,direction,rgb,3,0,0,5,23,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,52, 20,4,12); //back
+			} else {
+				buildFromImage(builder,direction,rgb,3,0,0,5,23,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,36, 48,4,4); //top
+				buildFromImage(builder,direction,rgb,3,0,0,5,12,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,40, 48,4,4); //bottom
+				buildFromImage(builder,direction,rgb,3,0,0,5,23,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,32, 52,4,12); //left
+				buildFromImage(builder,direction,rgb,0,0,0,8,23,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,40, 52,4,12); //right
+				buildFromImage(builder,direction,rgb,0,0,0,5,23,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,36, 52,4,12); //front
+				buildFromImage(builder,direction,rgb,3,0,0,8,23,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,44, 52,4,12); //back
 			}
 		}
-		
 		if(buildRight) {
-			//RIGHT ARM
-
-			//LEFT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftarmleft is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-xtimes,-5).setImageColor(img,48+xtimes,20+ytimes,rgb).queue(); //leftarmleft starts at 48, 20
-				} //leftarmleft starts
-			}
-
-			//RIGHT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftarmright is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-3+xtimes,-8).setImageColor(img,40+xtimes,20+ytimes,rgb).queue(); //leftarmright starts at 40, 20
-				} //leftarmright starts
-			}
-			
-		
-
-			//TOP
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //shoulderright is 4x4
-					Single.setup(builder,world,x,y+23,z).offset(direction,-3+ytimes,-5-xtimes).setImageColor(img,44+xtimes,16+ytimes,rgb).queue(); //shoulderright starts at 44, 16
-				} //shoulderright starts
-			}
-	
-			//BOTTOM
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //fistright is 4x4
-					Single.setup(builder,world,x,y+12,z).offset(direction,-3+ytimes,-5-xtimes).setImageColor(img,48+xtimes,16+ytimes,rgb).queue(); //fistright starts at 44, 16
-				} //fistright starts
-			}
-			
-			//FRONT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftarmfront is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,0,-8+xtimes).setImageColor(img,44+xtimes,20+ytimes,rgb).queue(); //leftarmfront starts at 44, 20
-				} //leftarmfront starts
-			}
-
-			//BACK
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftarmback is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-3,-5-xtimes).setImageColor(img,52+xtimes,20+ytimes,rgb).queue(); //leftarmback starts at 52, 20
-				} //leftarmback starts 
-			}
+			buildFromImage(builder,direction,rgb,3,0,7,0,23,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,44, 16,4,4); //top
+			buildFromImage(builder,direction,rgb,3,0,7,0,12,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,48, 16,4,4); //bottom
+			buildFromImage(builder,direction,rgb,3,0,7,0,23,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,40, 20,4,12); //left
+			buildFromImage(builder,direction,rgb,0,0,4,0,23,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,48, 20,4,12); //right
+			buildFromImage(builder,direction,rgb,0,0,7,0,23,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,44, 20,4,12); //front
+			buildFromImage(builder,direction,rgb,3,0,4,0,23,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,52, 20,4,12); //back
 		}
 	}
 
-	private static void buildArmsSlim(Builder builder, Level world, int x, int y, int z, BufferedImage img, Direction direction, boolean buildLeft, boolean buildRight, boolean rgb) {
+	private static void buildArmsSlim(Builder builder, BufferedImage img, Direction direction, boolean buildLeft, boolean buildRight, boolean rgb) {
 		if(buildLeft) {
-			//LEFT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightarmleft is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-xtimes,+4).setImageColor(img,47+xtimes,20+ytimes,rgb).queue();
-				} //rightarmleft starts at 47, 20
-			}
-
-			//RIGHT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightarmright is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-3+xtimes,+6).setImageColor(img,40+xtimes,20+ytimes,rgb).queue();
-				} //rightarmright starts at 40, 20
-			}
-
-			//TOP
-			for(int xtimes=0; xtimes<3; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //shoulderleft is 3x4
-					Single.setup(builder,world,x,y+23,z).offset(direction,-3+ytimes, 4 +xtimes).setImageColor(img,44+xtimes,16+ytimes,rgb).queue();
-				} //shoulderleft starts at 44, 16
-			}
-
-			//BOTTOM
-			for(int xtimes=0; xtimes<3; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //fistleft is 3x4
-					Single.setup(builder,world,x,y+12,z).offset(direction,-3+ytimes, 4 +xtimes).setImageColor(img,47+xtimes,16+ytimes,rgb).queue();
-				} //fistleft starts at 47, 16
-			}
-
-			//FRONT
-			for(int xtimes=0; xtimes<3; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightarmfront is 3x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,0, 6 -xtimes).setImageColor(img,44+xtimes,20+ytimes,rgb).queue();
-				} //rightarmfront starts at 44, 20 DONE?
-			}
-
-			//BACK
-			for(int xtimes=0; xtimes<3; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightarmback is 3x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-3, 4 +xtimes).setImageColor(img,51+xtimes,20+ytimes,rgb).queue();
-				} //rightarmback starts at 51, 20
+			if(img.getHeight() == 32) {
+				buildFromImage(builder,direction,rgb,3,0,0,7,23,0, RelativeDirection.LEFT, RelativeDirection.BACK,img,44, 16,3,4); //top
+				buildFromImage(builder,direction,rgb,3,0,0,7,12,0, RelativeDirection.LEFT, RelativeDirection.BACK,img,47, 16,3,4); //bottom
+				buildFromImage(builder,direction,rgb,0,0,0,5,23,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,47, 20,4,12); //left
+				buildFromImage(builder,direction,rgb,3,0,0,7,23,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,40, 20,4,12); //right
+				buildFromImage(builder,direction,rgb,0,0,0,7,23,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,44, 20,3,12); //front
+				buildFromImage(builder,direction,rgb,3,0,0,5,23,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,51, 20,3,12); //back
+			} else {
+				buildFromImage(builder,direction,rgb,3,0,0,5,23,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,36, 48,3,4); //top
+				buildFromImage(builder,direction,rgb,3,0,0,5,12,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,39, 48,3,4); //bottom
+				buildFromImage(builder,direction,rgb,3,0,0,5,23,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,32, 52,4,12); //left
+				buildFromImage(builder,direction,rgb,0,0,0,7,23,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,39, 52,4,12); //right
+				buildFromImage(builder,direction,rgb,0,0,0,5,23,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,36, 52,3,12); //front
+				buildFromImage(builder,direction,rgb,3,0,0,7,23,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,43, 52,3,12); //back
 			}
 		}
-
 		if(buildRight) {
-			//LEFT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftarmleft is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-xtimes,-5).setImageColor(img,48+xtimes,20+ytimes,rgb).queue();
-				} //leftarmleft starts at 48, 20
-			}
-
-			//RIGHT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftarmright is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-3+xtimes,-7).setImageColor(img,40+xtimes,20+ytimes,rgb).queue();
-				} //leftarmright starts at 40, 20
-			}
-
-			//TOP
-			for(int xtimes=0; xtimes<3; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //shoulderright is 4x4
-					Single.setup(builder,world,x,y+23,z).offset(direction,-3+ytimes,-5-xtimes).setImageColor(img,44+xtimes,16+ytimes,rgb).queue();
-				} //shoulderright starts at 44, 16
-			}
-
-			//BOTTOM
-			for(int xtimes=0; xtimes<3; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //fistright is 4x4
-					Single.setup(builder,world,x,y+12,z).offset(direction,-3+ytimes,-5-xtimes).setImageColor(img,47+xtimes,16+ytimes,rgb).queue();
-				} //fistright starts at 44, 16
-			}
-
-			//FRONT
-			for(int xtimes=0; xtimes<3; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftarmfront is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,0,-7+xtimes).setImageColor(img,44+xtimes,20+ytimes,rgb).queue();
-				} //leftarmfront starts at 44, 20
-			}
-
-			//BACK
-			for(int xtimes=0; xtimes<3; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftarmback is 4x12
-					Single.setup(builder,world,x,y+23-ytimes,z).offset(direction,-3,-5-xtimes).setImageColor(img,51+xtimes,20+ytimes,rgb).queue();
-				} //leftarmback starts at 52, 20
-			}
+			buildFromImage(builder,direction,rgb,3,0,6,0,23,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,44, 16,3,4); //top
+			buildFromImage(builder,direction,rgb,3,0,6,0,12,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,47, 16,3,4); //bottom
+			buildFromImage(builder,direction,rgb,3,0,6,0,23,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,40, 20,4,12); //left
+			buildFromImage(builder,direction,rgb,0,0,4,0,23,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,47, 20,4,12); //right
+			buildFromImage(builder,direction,rgb,0,0,6,0,23,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,44, 20,3,12); //front
+			buildFromImage(builder,direction,rgb,3,0,4,0,23,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,51, 20,3,12); //back
 		}
 	}
 	
-	private static void buildLegs(Builder builder, Level world, int x, int y, int z, BufferedImage img, Direction direction, boolean buildLeft, boolean buildRight, boolean rgb) {
+	private static void buildLegs(Builder builder, BufferedImage img, Direction direction, boolean buildLeft, boolean buildRight, boolean rgb) {
 		if(buildLeft) {
-			//LEFT LEG
-
-			//TOP
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //leftlegtop is 4x4
-					Single.setup(builder,world,x,y+11,z).offset(direction,-3+ytimes,+xtimes).setImageColor(img,4+xtimes,16+ytimes,rgb).queue(); //leftlegtop starts at 4, 16
-				} //leftlegtop starts at x, y+11, z-3
-			}
-
-			//BOTTOM
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //leftlegtop is 4x4
-					Single.setup(builder,world,x,y,z).offset(direction,-3+ytimes,+xtimes).setImageColor(img,8+xtimes,16+ytimes,rgb).queue(); //leftlegbot starts at 8, 16
-				} //leftlegbot starts at x, y, z-3
-			}
-
-			//LEFT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftlegleft is 4x12
-					Single.setup(builder,world,x,y+11-ytimes,z).offset(direction,-xtimes,+3).setImageColor(img,8+xtimes,20+ytimes,rgb).queue(); //leftlegleft starts at 8, 20
-				} //leftlegleft starts at x+3, y+11, z
-			}
-
-			//RIGHT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftlegright is 4x12
-					Single.setup(builder,world,x,y+11-ytimes,z).offset(direction,-3+xtimes,0).setImageColor(img, xtimes,20+ytimes,rgb).queue(); //leftlegright starts at 0, 20
-				} //leftlegright starts at x, y+11, z-3
-			}
-
-			//FRONT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftlegfront is 4x12
-					Single.setup(builder,world,x,y+11-ytimes,z).offset(direction,0,+xtimes).setImageColor(img,4+xtimes,20+ytimes,rgb).queue(); //leftlegfront starts at 4, 20
-				} //leftlegfront starts at x, y+11, z
-			}
-
-			//BACK
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //leftlegback is 4x12
-					Single.setup(builder,world,x,y+11-ytimes,z).offset(direction,-3, 3 -xtimes).setImageColor(img,12+xtimes,20+ytimes,rgb).queue(); //leftlegback starts at 12, 20
-				} //leftlegback starts at x+3, y+11, z+3
+			if(img.getHeight() == 32) {
+				buildFromImage(builder,direction,rgb,3,0,0,4,11,0, RelativeDirection.LEFT, RelativeDirection.BACK,img,4, 16,4,4); //top
+				buildFromImage(builder,direction,rgb,3,0,0,4,0,0, RelativeDirection.LEFT, RelativeDirection.BACK,img,8, 16,4,4); //bottom
+				buildFromImage(builder,direction,rgb,0,0,0,1,11,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,8, 20,4,12); //left
+				buildFromImage(builder,direction,rgb,3,0,0,4,11,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,0, 20,4,12); //right
+				buildFromImage(builder,direction,rgb,0,0,0,4,11,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,4, 20,4,12); //front
+				buildFromImage(builder,direction,rgb,3,0,0,1,11,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,12, 20,4,12); //back
+			} else {
+				buildFromImage(builder,direction,rgb,3,0,0,1,11,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,20, 48,4,4); //top
+				buildFromImage(builder,direction,rgb,3,0,0,1,0,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,24, 48,4,4); //bottom
+				buildFromImage(builder,direction,rgb,3,0,0,1,11,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,16, 52,4,12); //left
+				buildFromImage(builder,direction,rgb,0,0,0,4,11,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,24, 52,4,12); //right
+				buildFromImage(builder,direction,rgb,0,0,0,1,11,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,20, 52,4,12); //front
+				buildFromImage(builder,direction,rgb,3,0,0,4,11,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,28, 52,4,12); //back
 			}
 		}
-
 		if(buildRight) {
-			//RIGHT LEG
+			buildFromImage(builder,direction,rgb,3,0,3,0,11,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,4, 16,4,4); //top
+			buildFromImage(builder,direction,rgb,3,0,3,0,0,0, RelativeDirection.RIGHT, RelativeDirection.BACK,img,8, 16,4,4); //bottom
+			buildFromImage(builder,direction,rgb,3,0,3,0,11,0, RelativeDirection.BACK, RelativeDirection.DOWN,img,0, 20,4,12); //left
+			buildFromImage(builder,direction,rgb,0,0,0,0,11,0, RelativeDirection.FORWARD, RelativeDirection.DOWN,img,8, 20,4,12); //right
+			buildFromImage(builder,direction,rgb,0,0,3,0,11,0, RelativeDirection.RIGHT, RelativeDirection.DOWN,img,4, 20,4,12); //front
+			buildFromImage(builder,direction,rgb,3,0,0,0,11,0, RelativeDirection.LEFT, RelativeDirection.DOWN,img,12, 20,4,12); //back
+		}
+	}
 
-			//TOP
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //rightlegtop is 4x4
-					Single.setup(builder,world,x,y+11,z).offset(direction,-3+ytimes,-4+xtimes).setImageColor(img,4+xtimes,16+ytimes,rgb).queue(); //rightlegtop starts at 4, 16
-				} //rightlegtop starts at x-4, y+11, z-3
-			}
+	private enum RelativeDirection { LEFT,RIGHT,FORWARD,BACK,UP,DOWN }
 
-			//BOTTOM
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<4; ytimes++) { //rightlegbot is 4x4
-					Single.setup(builder,world,x,y,z).offset(direction,-3+ytimes,-4+xtimes).setImageColor(img,8+xtimes,16+ytimes,rgb).queue(); //rightlegbot starts at 8, 16
-				} //rightlegbot starts at x-4, y, z-3
-			}
-
-			//LEFT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightlegleft is 4x12
-					Single.setup(builder,world,x,y+11-ytimes,z).offset(direction,-xtimes,-1).setImageColor(img,8+xtimes,20+ytimes,rgb).queue(); //rightlegleft starts at 8, 20
-				} //rightlegleft starts at x+3, y+11, z
-			}
-
-			//RIGHT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightlegright is 4x12
-					Single.setup(builder,world,x,y+11-ytimes,z).offset(direction,-3+xtimes,-4).setImageColor(img, xtimes,20+ytimes,rgb).queue(); //rightlegright starts at 0, 20
-				} //rightlegright starts at x, y+11, z-3
-			}
-
-			//FRONT
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightlegfront is 4x12
-					Single.setup(builder,world,x,y+11-ytimes,z).offset(direction,0,-1-xtimes).setImageColor(img,4+xtimes,20+ytimes,rgb).queue(); //rightlegfront starts at 4, 20
-				} //rightlegfront starts at x, y+11, z
-			}
-
-			//BACK
-			for(int xtimes=0; xtimes<4; xtimes++) {
-				for(int ytimes=0; ytimes<12; ytimes++) { //rightlegback is 4x12
-					Single.setup(builder,world,x,y+11-ytimes,z).offset(direction,-3,-4+xtimes).setImageColor(img,12+xtimes,20+ytimes,rgb).queue(); //rightlegback starts at 12, 20
-				} //rightlegback starts at x+3, y+11, z+3
+	private static void buildFromImage(Builder builder, Direction direction, boolean rgb, int forward, int back, int left, int right, int up, int down, RelativeDirection dirX, RelativeDirection dirY, BufferedImage img, int imgX, int imgY, int sizeX, int sizeY) {
+		BlockPos pos = builder.getOriginPos();
+		for(int xtimes=0; xtimes<sizeX; xtimes++) {
+			for(int ytimes=0; ytimes<sizeY; ytimes++) {
+				int f = forward;
+				int b = back;
+				int l = left;
+				int r = right;
+				int u = up;
+				int d = down;
+				switch(dirX) {
+					case LEFT -> l += xtimes;
+					case RIGHT -> r += xtimes;
+					case FORWARD -> f += xtimes;
+					case BACK -> b += xtimes;
+					case UP -> u += xtimes;
+					case DOWN -> d += xtimes;
+				}
+				switch(dirY) {
+					case LEFT -> l += ytimes;
+					case RIGHT -> r += ytimes;
+					case FORWARD -> f += ytimes;
+					case BACK -> b += ytimes;
+					case UP -> u += ytimes;
+					case DOWN -> d += ytimes;
+				}
+				Single.setup(builder,builder.getWorld(),pos).offset(direction,f,b,l,r,u,d).setImageColor(img,imgX+xtimes,imgY+ytimes,rgb).queue();
 			}
 		}
 	}